# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CategoryEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryDTO": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OpnameResult": {
				"enum": [
					"match",
					"not_match"
				],
				"type": "string"
			},
			"OpnameEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"result": {
						"$ref": "#/components/schemas/OpnameResult"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/OpnameItemEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"code",
					"date",
					"result",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OpnameItemEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"paper_id": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"quantity_system": {
						"type": "number",
						"format": "double"
					},
					"opname": {
						"$ref": "#/components/schemas/OpnameEntity"
					}
				},
				"required": [
					"id",
					"paper_id",
					"quantity",
					"quantity_system"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OpnameItemDTO": {
				"properties": {
					"paper_id": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"paper_id",
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OpnameDTO": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/OpnameItemDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaperEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"stock": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"$ref": "#/components/schemas/CategoryEntity"
					},
					"opname_items": {
						"items": {
							"$ref": "#/components/schemas/OpnameItemEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"code",
					"stock",
					"category"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaperDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"stock": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"$ref": "#/components/schemas/CategoryEntity"
					}
				},
				"required": [
					"name",
					"stock",
					"category"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReportDetailEntity": {
				"properties": {
					"in": {
						"type": "number",
						"format": "double"
					},
					"out": {
						"type": "number",
						"format": "double"
					},
					"opname_code": {
						"type": "string"
					},
					"opname_date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"in",
					"out",
					"opname_code",
					"opname_date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReportEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"paper_code": {
						"type": "string"
					},
					"paper_name": {
						"type": "string"
					},
					"total_in": {
						"type": "number",
						"format": "double"
					},
					"total_out": {
						"type": "number",
						"format": "double"
					},
					"total_quantity": {
						"type": "number",
						"format": "double"
					},
					"category_name": {
						"type": "string"
					},
					"category_code": {
						"type": "string"
					},
					"category_id": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ReportDetailEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"paper_code",
					"paper_name",
					"total_in",
					"total_out",
					"total_quantity",
					"category_name",
					"category_code",
					"category_id",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ariestya",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "regyanax"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/category": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CategoryEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/category/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Show",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/opname": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OpnameEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Stock Opname"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpnameEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Stock Opname"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OpnameDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Stock Opname"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/opname/{id}/check": {
			"put": {
				"operationId": "Check",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpnameEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Stock Opname"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/opname/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpnameEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Stock Opname"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OpnameDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Show",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpnameEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Stock Opname"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/paper": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PaperEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Papers"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaperEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Papers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PaperDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Papers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/paper/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaperEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Papers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PaperDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Show",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaperEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Papers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/report": {
			"get": {
				"operationId": "Report",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ReportEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Reports"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "date_start",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "date_end",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/report/{id}": {
			"get": {
				"operationId": "Show",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReportEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Reports"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}